+++ common/src/main/java/com/nb6868/onex/common/util/TokenUtils.java
@@ -21,21 +21,16 @@ public class TokenUtils {
      * @return jwt
      */
     public static String getToken(AuthProps.Config authConfig, Map<String, Object> params) {
-        if ("uuid".equalsIgnoreCase(authConfig.getTokenPolicy())) {
-            return IdUtil.simpleUUID();
-        } else {
-            JWT jwt = JWT.create()
-                    .setSubject(authConfig.getType())
-                    .setKey(authConfig.getTokenKey().getBytes());
-            if (authConfig.getTokenExpire() == -1) {
+        Date now = new Date();
+        JWT jwt = JWT.create()
+                .setSubject(authConfig.getType())
+                .setKey(authConfig.getTokenKey().getBytes())
+                .setIssuedAt(now)
+                .setNotBefore(now)
                 // jwt99年过期=永不过期
-                jwt.setExpiresAt(DateUtil.offsetMonth(new Date(), 12 * 99));
-            } else {
-                jwt.setExpiresAt(DateUtil.offsetSecond(new Date(), authConfig.getTokenExpire()));
-            }
-            params.forEach(jwt::setPayload);
-            return jwt.sign();
-        }
+                .setExpiresAt(authConfig.getTokenExpire() == -1 ? DateUtil.offsetMonth(now, 12 * 99) : DateUtil.offsetSecond(now, authConfig.getTokenExpire()));
+        params.forEach(jwt::setPayload);
+        return jwt.sign();
     }
 
 }