+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # CHANGELOG
 
+## [3.0.9] 2022.03.29
+- upgrade: upgrade springboot to 2.6.5
+- enhance: shiro enhance
+
 ## [3.0.8] 2022.03.28
 - new: add shiro into common
 - enhance: optimize AccessControl shiro filter, support empty value
 - upgrade: upgrade hutool to 5.8.0.M1
+- remove: JwtShiroFilter/JwtTokenFilter
 
 ## [3.0.7] 2022.03.25
 - enhance: remove WxErrorException from exception handler